fix directory structure remove app_externals
