--- app_externals/gcal/__init__.py
@@ -1,5 +1,0 @@
-from .events import (
-    EventCreator, EventUpdater, EventDeleter,
-    EventGetter, EventLister,
-)
-from .open_service import GcalManagerAbs
