+++ notion_df/model/entity/block.py
@@ -3,7 +3,7 @@
 # 용어 정리: "Block" 이라 하면 (Block, PageBlock, PageRow, ...) 를 모두 가리킨다
 from __future__ import annotations
 
-from notion_df.core import Entity, MutableField, Field, ValueInput_T, Value_T, Property
+from notion_df.model.core import Value_T, ValueInput_T, Entity, Field, MutableField, Property
 
 
 class BlockIdField(Field['Block', str, str]):
