+++ notion_df/model/entity/workspace.py
@@ -1,7 +1,7 @@
 import os
 
-from notion_df.core import Entity
-from notion_df.util import get_num_iterator
+from notion_df.model.core import Entity
+from notion_df.util.util import get_num_iterator
 
 _num_iterator = get_num_iterator()
 
@@ -10,6 +10,7 @@ NOTION_API_KEY = os.getenv('NOTION_API_KEY')
 
 class Workspace(Entity):
     def __init__(self):
+        super().__init__()
         self._pk = next(_num_iterator)
 
     @property
