+++ notion_df/model/field_index.py
@@ -4,7 +4,7 @@ from abc import ABCMeta, abstractmethod
 from collections import defaultdict
 from typing import Final, Mapping, TypeVar, Any, Iterable, Callable
 
-from notion_df.core import Entity_T, Value_T, Field, FieldEventListener
+from notion_df.model.core import Entity_T, Value_T, Field, FieldEventListener
 from notion_df.util.collection import DictView
 from notion_df.util.mixin import Resolvable
