+++ notion_df/resource/property.py
@@ -1,7 +1,7 @@
 from dataclasses import dataclass
 from datetime import datetime
 
-from notion_df.resource.resource import Resource
+from notion_df.resource.core import Resource
 
 
 @dataclass
